pyproject.toml
[tool.poetry]
name = "pysnowdrop"
version = "1.0.6"
description = "Python Platform for DSGE Modeling"
authors = ["Alexei Goumilevski <agoumilevski@hotmail.com>",
           "James Otterson <jjotterson@gmail.com>"]
maintainers = ["Alexei Goumilevski <Agoumilevski@hotmail.com>",
               "James Otterson <jjotterson@gmail.com>"]
license = "LGPL-3.0-or-later"
readme =  "README.md"
keywords = ["DSGE", "Macroeconomic Modeling"]
classifiers=[
    "Programming Language :: Python",
    "Natural Language :: English",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Topic :: Software Development"
]

packages = [
    { include = "snowdrop/src", format = ["sdist", "wheel"] }
]

include = [
    { path = "snowdrop/bin", format = ["sdist", "wheel"] },
    { path = "supplements/data", format = ["sdist", "wheel"] },
    { path = "supplements/models", format = ["sdist", "wheel"] },
    { path = "supplements/docs", format = ["sdist", "wheel"] },
    { path = "supplements/api_docs", format = ["sdist", "wheel"] }
]

exclude = ["snowdrop/src/troll", "supplements/data/IMFE"]

[tool.poetry.scripts]
test = "tests.test_toy_models:test"

[tool.poetry.urls]
Repository = "https://github.com/gumilevskij/Framework.git"

# List od dependencies
[tool.poetry.dependencies]
python =  "^3.10"
pandas =  "^2.1.4"
scipy =  "1.12.0"
numpy =  "^1.26.4"
matplotlib =  "3.10.0"
sympy =  "^1.12"
symengine =  "^0.11.0"
emcee =  "^3.1.6"
statsmodels =  "^0.14.0"
pydot =  "^2.0.0"
networkx =  "^3.2"
particles =  "^0.4"
pymcmcstat =  "^1.9.1"
PyPDF2 =  "^3.0.1"
corner =  "^2.2.2"
seaborn =  "^0.13.0"
numba =  "^0.59.0"
lark =  "^1.2.2"
graphviz =  "^0.20.3"
filterpy =  "^1.4.5"
pytest =  "^7.4.0"
pykalman  =  "^0.9.7"
pylatex =  "^1.4.2"
pdfkit = "^1.0.0"
xlsxwriter = "^3.2.0"
"ruamel.yaml" = "0.17.21"
dill = "^0.3.8"
PyYAML = "^6.0.1"

# List of optional dependencies
compecon = {version="^2024.5.19", optional=true}
quantecon = {version="^0.7.2", optional=true}
reportlab = {version="^4.2.2", optional=true}
textwrap3 = {version="^0.9.2", optional=true}
PyLaTeX = {version="^1.4.2", optional=true}
glob2 = {version="^0.7", optional=true}
autograd = {version="^1.7.0", optional=true}
pdflatex =  {version="^0.1.3", optional=true}

[tool.poetry.extras]
reportlab =  ["reportlab"]
scikit-learn =  ["scikit-learn"]
pylatex =  ["PyLaTeX"]
textwrap3 =  ["textwrap3"]
glob2 =  ["glob2"]
pykalman  = ["pykalman"]
compecon = ["compecon"]
quantecon = ["quantecon"]
autograd = ["autograd"]
pdflatex = ["pdflatex"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
